tests/wpt/mozilla/tests/mozilla/bluetooth/getDescriptors/get-same-descriptors.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.heart_rate);
    return window.navigator.bluetooth.requestDevice({
        filters: [{services: [heart_rate.name]}],
        optionalServices: [generic_access.name]
    })
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService(generic_access.name))
    .then(service => service.getCharacteristic(device_name.name))
    .then(characteristic => Promise.all([
        characteristic.getDescriptor(number_of_digitals.name),
        characteristic.getDescriptor(number_of_digitals.name)]))
    .then(descriptors => {
        let descs1 = descriptors[0];
        let descs2 = descriptors[1];
        assert_equals(descs1.length, descs2.length);
        for (let i = 1; i < descs1.length; i++) {
            assert_equals(descs1[i], descs2[i]);
        }
    })
    .catch(error => assert_unreached('Promise should have resolved. ' + error));
}, 'Calls to get the same descriptor should return the same object.');
</script>
